fix doc string

+++ src/btk_pipeline/lib/window_stats.py
@@ -10,7 +10,7 @@ Options:
     --in TSV                 chunked summary stats tsv file.
     --min-window-length INT  minimum length of a window. [Default: 100000]
     --min-window-count INT   minimum number of windows. [Default: 1]
-    --window INT             window size or proportion. [Default: 1]
+    --window FLOAT           window size or proportion. [Default: 1]
     --out TSV                output TSV filename or suffix.
 """
